Create 'get_dictionary_of_food_items' function
The function serves as a context object for the recordset.html template.
It returns a dictionary where food categories are the keys and currently (non)registered food items associated with that category are the values
All FoodItem objects known to the freezer registration system are included if their attribute 'currently_registered' is set to the same boolen value as parameter 'show_currently_registered'

With such dictionary, the recordset.html template will then be able to list all (non)registered items grouped by their categories.

The Recordset view class now has the 'context_object_name' set to this dictionary variable name, which is 'categories_items_dictionary.