Create for loop in recordset.html template for listing all food categories and their items passed in context object
The main purpose of recordset.html template is to list all categories and their associated items currently (non)registered in the freezer.
The for loop also creates some very basic control interface for maniuplating with the amount of listed items.
Basic css class names created during the proccess that will be later on used to style the elements.