Create 'record-items.html'template
this template will be the opposite of the recordset.html template.

While 'recordset.html' displays all the items currently listed in a freezer, the 'record-items.html' template will list all the items, that are known to this system, but not currently recorded in a freezer.

The idea is that if the user removes some type of item from the freezer, it is still being held in a system as 'often recorded'.

On a 'record-items.html' template user will have a quick and easy option to record back any item that previously was recorded.

At the same time, the user will have the option to remove items permanently from the 'record-items.html' template. After that, they will no longer be available for quick record. If such items is ever to be recorded again, it will need to be registered from scratch to the system.