Create 'register-items.html' template
Here, the new FoodItems are about to be registered in the system.
A simple Django Form will do the job.

Newly registered items will then appear in the FoodItem database, but still as currently non-registered
They appear in the system, but if they are agbout to be listed in the recordset, they need to be manually added by user thorugh 'record-items.html' template